--- discord/utils.go
@@ -1,16 +1,0 @@
-package discord
-
-import (
-	"math/rand"
-	"time"
-)
-
-func randomString() string {
-	var seededRand = rand.New(rand.NewSource(time.Now().UnixNano()))
-	charset := "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0987654321"
-	b := make([]byte, 20)
-	for i := range b {
-		b[i] = charset[seededRand.Intn(len(charset)-1)]
-	}
-	return string(b)
-}
